Modified the method to check stop/start cmds
Modified the method to check stop/start commands from $? to the 'outcome' string from the jboss-cli.sh exit. This is because, when I ask if I could use $? to check the status of the stop/start command to a Red Hat analyst, he told me "As the engineering team said, when using pure CLI, you would look for the "outcome" field to contain "success" or "failure". It would probably be the best way to make sure the command was issued successfuly."